+++ src/main/java/com/forgestove/create_cyber_goggles/event/ModKeyMapping.java
@@ -8,12 +8,12 @@ import org.jetbrains.annotations.NotNull;
 
 import static com.forgestove.create_cyber_goggles.CreateCyberGoggles.ID;
 import static com.mojang.blaze3d.platform.InputConstants.Type.KEYSYM;
-import static net.neoforged.neoforge.client.settings.KeyConflictContext.IN_GAME;
+import static net.neoforged.neoforge.client.settings.KeyConflictContext.*;
 import static org.lwjgl.glfw.GLFW.GLFW_KEY_UNKNOWN;
 public class ModKeyMapping {
 	public static final Lazy<KeyMapping> OPEN_CONFIG = register("openConfig", IN_GAME, KEYSYM, GLFW_KEY_UNKNOWN);
 	public static final Lazy<KeyMapping> OPEN_STOCK = register("openStock", IN_GAME, KEYSYM, GLFW_KEY_UNKNOWN);
-	public static final Lazy<KeyMapping> PREVIEW_FILTER = register("previewFilter", IN_GAME, KEYSYM, GLFW_KEY_UNKNOWN);
+	public static final Lazy<KeyMapping> PREVIEW_FILTER = register("previewFilter", UNIVERSAL, KEYSYM, GLFW_KEY_UNKNOWN);
 	public static final Lazy<KeyMapping> TOGGLE_DIVING = register("toggleDiving", IN_GAME, KEYSYM, GLFW_KEY_UNKNOWN);
 	@SuppressWarnings("SameParameterValue")
 	private static @NotNull Lazy<KeyMapping> register(String name, KeyConflictContext context, Type type, int key) {